% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bmass_TestData2}
\alias{bmass_TestData2}
\title{bmass Example Dataset 2}
\format{A data frame with 11 rows and 9 variables:
\describe{
  \item{Chr}{chromosome}
  \item{BP}{basepair position}
  \item{Marker}{rsID# or other identifier}
  \item{MAF}{Minor Allele Frequency}
  \item{A1}{reference allele}
  \item{A2}{alternative allele}
  \item{Direction}{direction of association
effect size, + or -}
  \item{pValue}{p-Value of GWAS association}
  \item{N}{sample size}
}}
\source{
Manually created
}
\usage{
bmass_TestData2
}
\description{
A manually created sample dataset for use in vignettes
and unit tests
}
\keyword{datasets}
