% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bmass_TestSigSNPs}
\alias{bmass_TestSigSNPs}
\title{bmass Example GWAS SNPs}
\format{A data frame with 2 rows and 2 variables:

\describe{
  \item{Chr}{chromosome}
  \item{BP}{basepair position}
}}
\source{
Manually created
}
\description{
A manually created list of GWAS significant SNPs to be
used in conjunction with 'bmass_TestData1' and 'bmass_TestData2'.
}
\keyword{data}
