% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetResultsFromMarginalSNPsAndFormat.R
\name{CollapseSigmaAlphasTogether}
\alias{CollapseSigmaAlphasTogether}
\title{marginal' of across all sigma_alphas)}
\usage{
CollapseSigmaAlphasTogether(inputValues1, nSigmaAlphas)
}
\description{
marginal' of across all sigma_alphas)
}
