% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetResultsFromMarginalSNPsAndFormat.R
\name{DetermineAndApplyPriors}
\alias{DetermineAndApplyPriors}
\title{DetermineAndApplyPriors}
\usage{
DetermineAndApplyPriors(DataSources, MarginalSNPs, GWASsnps, SigmaAlphas,
  Models, ModelPriors, ProvidedPriors, UseFlatPriors, GWASThreshFlag,
  GWASThreshValue, bmassSeedValue, LogFile)
}
\description{
DetermineAndApplyPriors
}
\keyword{internal}
