% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeDataGWASAnnotateAndGetMarginalSNPs.R
\name{GetZScoreAndDirection}
\alias{GetZScoreAndDirection}
\title{GetZScoreAndDirection}
\usage{
GetZScoreAndDirection(DataSources1)
}
\description{
GetZScoreAndDirection
}
\keyword{internal}
