% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeDataGWASAnnotateAndGetMarginalSNPs.R
\name{ProcessMergedAndAnnotatedDataSources}
\alias{ProcessMergedAndAnnotatedDataSources}
\title{ProcessMergedAndAnnotatedDataSources}
\usage{
ProcessMergedAndAnnotatedDataSources(DataSources, MergedDataSources,
  ZScoresCorMatrix, SNPMarginalUnivariateThreshold,
  SNPMarginalMultivariateThreshold, LogFile)
}
\description{
ProcessMergedAndAnnotatedDataSources
}
\keyword{internal}
