% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmass.R
\name{bmass}
\alias{bmass}
\title{Bayesian multivariate analysis of summary statistics (bmass)}
\usage{
bmass <- function (DataSources, GWASsnps = NULL, GWASThreshFlag = FALSE, GWASThreshValue = 5e-8, NminThreshold = 0)
}
\arguments{
\item{DataSources}{A string indicating the variable names of the input datafiles and phenotypes}

\item{GWASsnps}{A data.table containing rows of SNPs that were univariate genome-wide significant in the phenotypes being used for analysis; GWASsnps input file should have two columns, one for chromosome and another for basepair position (with column headers of "Chr" and "BP")}

\item{GWASThreshFlag}{A logical TRUE/FALSE flag that indicates whether to threshold input GWASsnps list by a univariate GWAS p-value or not (eg the input GWASsnps list contains variants that are significant from discovery + replication data, but the input summary statistics are just from the discovery cohort). Default is FALSE.}

\item{GWASThreshValue}{A numerical value indicating the univariate p-value threshold to use in conjunction with the GWASThreshFlag. Default is 5e-8.}

\item{NminThreshold}{A numerical value that indicates a sample size threshold to use where SNPs below which are removed. Default is 0.}

\item{PrintProgress}{A logical TRUE/FALSE flag that indicates whether progress statements should be printed to stderr() during the course of running bmass() or not. Default is FALSE.}
}
\value{
A list containing model, SNP, and posterior information for both the previously significant univariate SNPs ("PreviousSNPs") and the newly significant multivariate SNPs ("NewSNPs"). For a full breakdown of the bmass() output list structure, please see the associated README and/or vignettes.
}
\description{
Run bmass on a set of phenotypes that each have univariate GWAS statistics on the same set of SNPs
}
\examples{
bmass(c("HDL", "LDL", "TG", "TC"), GWASsnps=InputGWASSNPsFile, NminThreshold = 50000) 
bmass(c("HDL", "LDL", "TG", "TC"), GWASsnps=InputGWASSNPsFile, GWASThreshFlag = FALSE, GWASThreshValue = 1e-8, NminThreshold = 50000, PrintProgress=TRUE) 
bmassOutput <- list(); bmassOutput <- bmass(c("HDL", "LDL", "TG", "TC"), GWASsnps=InputGWASSNPsFile, NminThreshold = 50000) 

}

